完成产品加工添加功能

+++ src/main/java/com/happycode/model/OrderDetailSummary.java
@@ -7,10 +7,10 @@ import java.math.BigDecimal;
 @Data
 public class OrderDetailSummary {
     private String productName;
-    private Long quantity;
+    private double quantity;
     private BigDecimal unitPrice;
 
-    public OrderDetailSummary(String productName, Long quantity, BigDecimal unitPrice) {
+    public OrderDetailSummary(String productName, double quantity, BigDecimal unitPrice) {
         this.productName = productName;
         this.quantity = quantity;
         this.unitPrice = unitPrice;
